backend/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis[hiredis]==5.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation & Config
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# AI
google-generativeai>=0.5.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0

# Utilities
email-validator==2.1.0.post1
